Move the relevant functions to the 'ocs_ci/utility/rosa.py' file, Add tier marker, fix get pod names
