+++ tests/manage/z_cluster/nodes/test_nodes_restart_ms.py
@@ -303,6 +303,7 @@ class TestNodesRestartMS(ManageTest):
         # Create PVCs and pods
         self.sanity_helpers.create_resources_on_ms_consumers()
 
+    @tier4a
     @polarion_id("OCS-3982")
     def test_provider_worker_nodes_restart(self, nodes):
         """
@@ -333,6 +334,7 @@ class TestNodesRestartMS(ManageTest):
         verify_provider_osd_nodes_on_correct_machine_pools()
         self.sanity_helpers.create_resources_on_ms_consumers()
 
+    @tier4b
     @pytest.mark.polarion_id("OCS-2015")
     def test_rolling_provider_worker_nodes_restart(self, nodes):
         """
